requirements.txt
opencv-python==4.9.0.80
python-opencv-utils==0.0.4
bootstrap-py==2.0.0
numpy==1.26.4
matplotlib==3.8.3
kafka-python==2.0.2
redis==5.0.1
pyspark==3.5.0
yolov5==7.0.13
requests==2.31.0
fastapi==0.109.2
fastapi-utils==0.2.1
fastapi-frame-stream==0.1.1
python-multipart==0.0.9
pydantic>=1.2.0, <2.0.0
httpx==0.26.0
uvicorn==0.27.0.post1